requirements.txt
Pillow
matplotlib
torch==1.6.0
torchvision
tensorboard